requirements.txt
python-ceilometerclient